# api/swagger.yaml
basePath: /beetle
definitions:
...
  map_string_string:
    additionalProperties:
      type: string
    type: object
  model.ApiResponse-any:
    properties:
      data:
        description: Contains the actual response data (single object, list, or page)
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
...
  model.ApiResponse-map_string_string:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/map_string_string'
        description: Contains the actual response data (single object, list, or page)
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
  model.ApiResponse-string:
    properties:
      data:
        description: Contains the actual response data (single object, list, or page)
        type: string
      error:
        description: Error message for failed responses
        example: Error message if failure
        type: string
      message:
        description: Optional message for additional context
        example: Operation successful
        type: string
      success:
        description: Indicates whether the API call was successful
        example: true
        type: boolean
    type: object
...
host: localhost:8056
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cm-beetle/issues/new/choose
  description: CM-Beetle recommends optimal target cloud infrastructure and executes
    migration.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Beetle API
  version: latest
paths:
...
  /test/auth:
    get:
      consumes:
      - application/json
      description: Checks if the request is authenticated and returns the auth type
      operationId: TestAuth
      produces:
      - application/json
      responses:
        "200":
          description: Successfully authenticated
          schema:
            $ref: '#/definitions/model.ApiResponse-map_string_string'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      security:
      - BasicAuth: []
      - BearerAuth: []
      summary: Test authentication
      tags:
      - '[Test] Utilities'
      x-order: 2
...
  /test/tracing:
    get:
      consumes:
      - application/json
      description: |-
        Tests distributed tracing by calling Tumblebug's readyz endpoint.

        [Note]
        - The 'traceparent' header (W3C Trace Context) is propagated to Tumblebug for distributed tracing.
        - The 'X-Request-Id' header is propagated for log correlation.
        - Use this API to verify that tracing works correctly between Beetle and Tumblebug.
      operationId: TestTracing
      parameters:
      - description: W3C Trace Context (e.g., 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01).
          Used for distributed tracing.
        in: header
        name: traceparent
        type: string
      - description: Unique request ID. Used for tracking request status and correlating
          logs.
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tumblebug service is ready
          schema:
            $ref: '#/definitions/model.ApiResponse-string'
        "503":
          description: Tumblebug service unavailable
          schema:
            $ref: '#/definitions/model.ApiResponse-any'
      tags:
      - '[Test] Utilities'
security:
- BasicAuth: []
- BearerAuth: []
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...